requirements.txt
mss
pillow
imagehash
rich>=13.0.0
google-generativeai>=0.8.3
pillow>=10.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pynput>=1.7.6
click